Move base class to base
