+++ src/pykeen/models/unimodal/complex.py
@@ -7,94 +7,16 @@ from typing import Optional
 import torch
 import torch.nn as nn
 
-from ..base import EntityRelationEmbeddingModel, InteractionFunction, normalize_for_einsum
+from ..base import InteractionFunction, SimpleVectorEntityRelationEmbeddingModel, normalize_for_einsum
 from ...losses import Loss, SoftplusLoss
 from ...regularizers import LpRegularizer, Regularizer
 from ...triples import TriplesFactory
-from ...utils import get_embedding_in_canonical_shape, split_complex
+from ...utils import split_complex
 
 __all__ = [
     'ComplEx',
     'ComplexInteractionFunction',
 ]
-
-
-class SimpleVectorEntityRelationEmbeddingModel(EntityRelationEmbeddingModel):
-    """A base class for embedding models which store a single vector for each entity and relation."""
-
-    def __init__(
-        self,
-        triples_factory: TriplesFactory,
-        interaction_function: InteractionFunction,
-        embedding_dim: int = 200,
-        automatic_memory_optimization: Optional[bool] = None,
-        loss: Optional[Loss] = None,
-        preferred_device: Optional[str] = None,
-        random_seed: Optional[int] = None,
-        regularizer: Optional[Regularizer] = None,
-    ) -> None:
-        """Initialize embedding model.
-
-        :param triples_factory: TriplesFactory
-            The triple factory connected to the model.
-        :param interaction_function:
-            The interaction function used to compute scores.
-        :param embedding_dim:
-            The embedding dimensionality of the entity embeddings.
-        :param automatic_memory_optimization: bool
-            Whether to automatically optimize the sub-batch size during training and batch size during evaluation with
-            regards to the hardware at hand.
-        :param loss: OptionalLoss (optional)
-            The loss to use.
-        :param preferred_device: str (optional)
-            The default device where to model is located.
-        :param random_seed: int (optional)
-            An optional random seed to set before the initialization of weights.
-        :param regularizer: BaseRegularizer
-            The regularizer to use.
-        """
-        super().__init__(
-            triples_factory=triples_factory,
-            embedding_dim=embedding_dim,
-            automatic_memory_optimization=automatic_memory_optimization,
-            loss=loss,
-            preferred_device=preferred_device,
-            random_seed=random_seed,
-            regularizer=regularizer,
-        )
-
-        self.interaction_function = interaction_function
-
-        # Finalize initialization
-        self.reset_parameters_()
-
-    def _score(
-        self,
-        h_ind: Optional[torch.LongTensor] = None,
-        r_ind: Optional[torch.LongTensor] = None,
-        t_ind: Optional[torch.LongTensor] = None,
-    ) -> torch.FloatTensor:
-        # Get embeddings
-        h = get_embedding_in_canonical_shape(embedding=self.entity_embeddings, ind=h_ind)
-        r = get_embedding_in_canonical_shape(embedding=self.relation_embeddings, ind=r_ind)
-        t = get_embedding_in_canonical_shape(embedding=self.entity_embeddings, ind=t_ind)
-
-        # Compute score
-        scores = self.interaction_function(h=h, r=r, t=t)
-
-        # Only regularize relation embeddings
-        self.regularize_if_necessary(r)
-
-        return scores
-
-    def score_hrt(self, hrt_batch: torch.LongTensor) -> torch.FloatTensor:  # noqa: D102
-        return self._score(h_ind=hrt_batch[:, 0], r_ind=hrt_batch[:, 1], t_ind=hrt_batch[:, 2]).view(-1, 1)
-
-    def score_t(self, hr_batch: torch.LongTensor) -> torch.FloatTensor:  # noqa: D102
-        return self._score(h_ind=hr_batch[:, 0], r_ind=hr_batch[:, 1], t_ind=None).view(-1, self.num_entities)
-
-    def score_h(self, rt_batch: torch.LongTensor) -> torch.FloatTensor:  # noqa: D102
-        return self._score(h_ind=None, r_ind=rt_batch[:, 0], t_ind=rt_batch[:, 1]).view(-1, self.num_entities)
 
 
 class ComplexInteractionFunction(InteractionFunction):
